Safari-611.1.10/LayoutTests/fast/canvas/webgl/texture-active-bind-expected.txt


Tests that glActiveTexture and glBindTexture work as expected
Specifically texture targets are per active texture unit.

PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS gl.getError() is gl.NO_ERROR
PASS expected:0, 192, 128, 255 found: 0, 192, 128, 255
PASS expected:128, 64, 255, 255 found: 128, 64, 255, 255
PASS expected:192, 255, 64, 255 found: 192, 255, 64, 255
PASS expected:200, 0, 255, 255 found: 200, 0, 255, 255
PASS successfullyParsed is true

TEST COMPLETE
